requirements.txt
# ── Web framework & server ─────────────────────────────────────
flask==2.3.3
werkzeug==2.3.7
gunicorn==21.2.0

# ── Database ───────────────────────────────────────────────────
flask-pymongo==2.3.0
pymongo==4.6.0

# ── Auth & forms ───────────────────────────────────────────────
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.2.1
wtforms==3.1.1

# ── Computer vision (headless = no display required on cloud) ──
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
mediapipe==0.10.8

# ── AI / ML ────────────────────────────────────────────────────
tensorflow==2.13.0
keras==2.13.1
tf-keras>=2.16.0
scikit-learn==1.3.2
deepface>=0.0.93

# ── External APIs ──────────────────────────────────────────────
replicate>=0.25.0
requests==2.31.0

# ── Environment ────────────────────────────────────────────────
python-dotenv==1.0.0
